// apps/frontend/src/client/models/IdentifyResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IdentifyResponse = {
    matched: boolean;
    person_id?: (string | null);
    name?: (string | null);
    document?: (string | null);
    score: number;
    distance: number;
    processing_time_ms?: (number | null);
};